+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### New Features
 
 - Added parameters `external_access_integrations` and `secrets` when creating a UDF, UDTF or Stored Procedure from Snowpark Python to allow integration with external access.
+- Added support for these new functions in `snowflake.snowpark.functions`:
+  - `array_flatten`
+- Added support for replicating your local Python environment on Snowflake via `Session.replicate_local_environment`.
 
 ## 1.6.1 (2023-08-02)